+++ NewsReader/Model/Categories.swift
@@ -47,7 +47,6 @@ class MakeNewsCategories {
     }
 }
 
-
 class NewsReaderData {
     var categories: [NewsCategories]?
     var favoriteNews: [NewsArticleModel]?
