+++ NewsReader/NewsReader/NewsReaderViewController.swift
@@ -34,41 +34,4 @@ class NewsReaderViewController: UIViewController {
 }
 
 
-extension NewsReaderViewController {
-    
-    func newsRequest(_ keyword: String) {
-        
-        parameters = ["q" : keyword,
-                      "pageSize" : pageSize,
-                      "page" : pageNumber]
-        
-        if !isLoading {
-            
-            let url = URL(string: "https://newsapi.org/v2/everything")
-            if let recieveUrl = url {
-                Alamofire.request(recieveUrl, method: .get,
-                                  parameters: parameters,
-                                  encoding: URLEncoding.default,
-                                  headers: ["X-Api-Key": "4ea21ee288f24ae880ef13ebda15edbd"]).responseData { (respomse) in
-                    self.isLoading = true
-                        if let date = respomse.result.value {
-                            do {
-                                let newsModel = try JSONDecoder().decode(NewsModel.self, from: date)
-                                    if let articles = newsModel.articles {
-                                        self.newsList.append(contentsOf: articles)
-                                    } else {
-                                        self.showErrorAlert("news not found")
-                                    }
-                                    self.collectionView.reloadData()
-                                } catch {
-                                    self.showErrorAlert("news not found or thms else")
-                                }
-                            }
-                }
-            } else {
-                showErrorAlert("fatal error")
-            }
-        }
-        
-    }
-}
+
